Invoker generified / mvn
